Machine_Learning-Image_Super_Resolution/requirements.txt
flask
flask-cors
waitress
tensorflow-cpu
